fix: load classic-parse worker from frontend url
The hardcoded /editor/scene worker path loaded the deployed worker regardless of config.url.frontend, so under a local or preview frontend the new source-text-protocol client ran against the stale url-protocol worker and every classic script parse failed. Resolve it from config.url.frontend like every other worker.

Also inline the single-use sandbox helpers into their call site.
